Move TransactionClient and aux structures to new module
